requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
gunicorn==21.2.0
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==20.5
# Removed unused dependencies like pymongo, flask, etc.
